Share adjust titleBar code for shell and navigationPage
